Chapter 09/training/requirements.txt
scikit-learn
pandas
polyaxon
feast[gcp]
google-cloud-bigquery
google-auth